Changes for FIT 2 implemented

Approved-by: Christian Eichler <code@christian-eichler.de>
Approved-by: kiozen <oliver.eichler@gmx.de>

+++ src/gis/fit/defs/CFitProfile.cpp
@@ -50,7 +50,7 @@ const CFitFieldProfile* CFitProfile::getField(quint8 fieldDefNr) const
 }
 
 
-void CFitProfile::addField(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, quint16 offset, QString units)
+void CFitProfile::addField(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, qint16 offset, QString units)
 {
     CFitFieldProfile* field= new CFitFieldProfile(this, name, baseType, fieldDefNr, scale, offset, units);
     fields.insert(fieldDefNr, field);
@@ -61,12 +61,12 @@ void CFitProfile::addSubfield(QString name, const CFitBaseType& baseType, quint8
     CFitSubfieldProfile* field = new CFitSubfieldProfile(this, name, baseType, fieldDefNr, scale, offset, units, subRefFieldDefNr, subRefFieldValue);
     fields[fieldDefNr]->addSubfield(field);
 }
-void CFitProfile::addComponent(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, quint16 offset, QString units, quint8 componentFieldDefNr, quint8 bits)
+void CFitProfile::addComponent(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, qint16 offset, QString units, quint8 componentFieldDefNr, quint32 bits)
 {
     CFitComponentfieldProfile* field = new CFitComponentfieldProfile(this, name, baseType, fieldDefNr, scale, offset, units, componentFieldDefNr, bits);
     fields[fieldDefNr]->addComponent(field);
 }
-void CFitProfile::addComponent(int subfieldIndex, QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, quint16 offset, QString units, quint8 componentFieldDefNr, quint8 bits)
+void CFitProfile::addComponent(int subfieldIndex, QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, qint16 offset, QString units, quint8 componentFieldDefNr, quint32 bits)
 {
     CFitComponentfieldProfile* field = new  CFitComponentfieldProfile(this, name, baseType, fieldDefNr, scale, offset, units, componentFieldDefNr, bits);
     fields[fieldDefNr]->addComponent(subfieldIndex, field);
